+++ src/container.ml
@@ -9,6 +9,7 @@ let with_return = With_return.with_return
 
 type ('t, 'a, 'accum) fold = 't -> init:'accum -> f:('accum -> 'a -> 'accum) -> 'accum
 type ('t, 'a) iter = 't -> f:('a -> unit) -> unit
+type 't length = 't -> int
 
 let iter ~fold t ~f = fold t ~init:() ~f:(fun () a -> f a)
 
@@ -82,7 +83,15 @@ let find ~iter c ~f =
 
 let to_list ~fold c = List.rev (fold c ~init:[] ~f:(fun acc x -> x :: acc))
 
-let to_array ~fold c = Array.of_list (to_list ~fold c)
+let to_array ~length ~iter c =
+  let array = ref [||] in
+  let i = ref 0 in
+  iter c ~f:(fun x ->
+    if !i = 0 then (array := Array.create ~len:(length c) x);
+    !array.(!i) <- x;
+    incr i);
+  !array
+;;
 
 module Make_gen (T : Make_gen_arg) : sig
   include Generic with type 'a t := 'a T.t with type 'a elt := 'a T.elt
@@ -95,7 +104,12 @@ end = struct
     | `Define_using_fold -> fun t ~f -> iter ~fold t ~f
   ;;
 
-  let length t       = length   ~fold t
+  let length =
+    match T.length with
+    | `Custom length     -> length
+    | `Define_using_fold -> fun t -> length ~fold t
+  ;;
+
   let is_empty t     = is_empty ~iter t
   let sum m t        = sum      ~fold m t
   let count t ~f     = count    ~fold t ~f
@@ -104,7 +118,7 @@ end = struct
   let find_map t ~f  = find_map ~iter t ~f
   let find t ~f      = find     ~iter t ~f
   let to_list t      = to_list  ~fold t
-  let to_array t     = to_array ~fold t
+  let to_array t     = to_array ~length ~iter t
 
   let min_elt t ~compare = min_elt  ~fold t ~compare
   let max_elt t ~compare = max_elt  ~fold t ~compare